requirements.txt
psycopg2
geojson
django==1.2.X
django-olwidget
